updated unlink

+++ app/Controllers/BugReportController.php
@@ -56,7 +56,10 @@ class BugReportController extends BaseController
             if ($file->isValid() && !$file->hasMoved()) {
                 $newName = $file->getRandomName();
                 $file->move('../public/uploads', $newName);
-                unlink("../public/uploads/" . $bug->reference);
+                if($bug->reference==""){
+                    unlink("../public/uploads/" . $bug->reference);
+                }
+
             }
         }
